Connection and OpenAPI & Fault Tolerance

+++ src/main/java/com/distribuida/db/Instrument.java
@@ -7,16 +7,16 @@ import jakarta.persistence.*;
 public class Instrument {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name")
     private String name;
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
